nsmgr: rearrange file structure
to decrease churn from moving packages, these patches were left out to keep the original patch
as small as possible.

This commit rearranges the nsmgr package more logically
